pyproject.toml
[project]
name = "workspaces3"
version = "0.1.0"
description = "Autonomous AI agent for task completion - Manus competitor"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workspaces3"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pyright]
venvPath = ".."
venv = ".venv"
pythonVersion = "3.11"
typeCheckingMode = "basic"
exclude = ["**/__pycache__", ".venv"]
